requirements.txt
bleach==6.1.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
coverage==7.13.0
exceptiongroup==1.3.1
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.3.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
Pillow>=10.1.0
pluggy==1.6.0
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
six==1.17.0
# Python 3.13 호환: 2.0.23은 TypingOnly/__static_attributes__ 오류 발생 → 2.0.36+ 사용
SQLAlchemy>=2.0.36
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.6.2
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.1
